PR-07: extract lifecycle/{mod,logind,startup,snapshot}.rs
Behaviour-preserving relocation of the lifecycle-provider cluster
into a directory module. ~640 LOC moved.

- lifecycle/mod.rs: LifecycleProvider enum + impl, plus pub(crate)
  re-exports of the three submodules so main.rs's lifecycle::*
  glob reaches submodule items (per plan D13).
- lifecycle/logind.rs: all logind parse/decode helpers,
  LogindLifecycleProvider, monitor task helpers, snapshot decoders.
- lifecycle/startup.rs: StartupSnapshotProvider.
- lifecycle/snapshot.rs: snapshot_no_session helper.

main.rs: 5579 -> 4911 LOC. resolve_logind_session_path widened
to pub(crate) because resolve_display_override_from_logind (still
in main.rs until PR-08) consumes it.

Verification: cargo build clean (warning count 1 -> 7, all unused
imports in main.rs from incomplete cleanup — follow-up tracked);
cargo test --bin kanata-switcher 261 passed / 0 failed.
